# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TripFit API",
    "version": "v0.0.1"
  },
  "servers": [
    {
      "url": "https://api.tripfit.online",
      "description": "Production API Server"
    }
  ],
  "security": [
    {
      "bearer-jwt": []
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "액세스 토큰 재발급",
        "description": "목적: refresh token으로 access JWT만 다시 발급한다.\n\n호출 시점: access 만료 직전·401 이후 재시도.\n\n전제: 아직 폐기되지 않은 유효한 refresh token.\n\n결과: 새 access JWT. refresh row는 유지된다.\n\n주요 에러: AUTH_INVALID_REFRESH — refresh 무효·만료\n",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "재발급 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseRefreshResponse"
                },
                "example": {
                  "data": {
                    "accessToken": "eyJhbG...",
                    "expiresIn": 7200
                  }
                }
              }
            }
          },
          "400": {
            "description": "요청 값 검증 실패 (INVALID_INPUT)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "INVALID_INPUT",
                  "message": "입력값이 올바르지 않습니다.",
                  "errors": [
                    {
                      "field": "refreshToken",
                      "message": "필수 값입니다."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "AUTH_INVALID_REFRESH — refresh 무효·만료",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_INVALID_REFRESH",
                  "message": "유효하지 않은 리프레시 토큰입니다."
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "로그아웃",
        "description": "목적: refresh token을 폐기해 재발급을 막는다.\n\n호출 시점: 사용자가 로그아웃아웃할 때.\n\n전제: 본인이 보유한 refresh token.\n\n결과: 204 No Content. access는 만료까지 유효할 수 있다.\n",
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "로그아웃 성공(No Content)"
          },
          "400": {
            "description": "요청 값 검증 실패 (INVALID_INPUT)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "INVALID_INPUT",
                  "message": "입력값이 올바르지 않습니다.",
                  "errors": [
                    {
                      "field": "refreshToken",
                      "message": "필수 값입니다."
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "소셜 로그인",
        "description": "목적: 소셜 토큰으로 로그인하고 access·refresh를 발급한다.\n\n호출 시점: 앱 최초 로그인·재로그인.\n\n전제: Google/Kakao/Apple에서 받은 유효한 토큰.\n\n결과: access·refresh 토큰과 사용자 요약(hasPreSchedule·isAllFree 포함).\n\n주요 에러: AUTH_SOCIAL_TOKEN_EXPIRED — 소셜 토큰 만료(재로그인 유도) · AUTH_SOCIAL_TOKEN_INVALID — 그 외 소셜 토큰 무효 · AUTH_SOCIAL_PROVIDER_UNAVAILABLE — 소셜 provider 접근 실패(재시도 유도)\n",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseLoginResponse"
                },
                "example": {
                  "data": {
                    "accessToken": "eyJhbG...",
                    "refreshToken": "550e8400-e29b-41d4-a716-446655440000",
                    "expiresIn": 7200,
                    "user": {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "email": "user@example.com",
                      "firstName": "길동",
                      "lastName": "홍",
                      "nickname": "홍길동",
                      "profileImageUrl": "https://lh3.googleusercontent.com/a/example",
                      "provider": "GOOGLE",
                      "isGoogleCalendarConnected": false,
                      "hasPreSchedule": false,
                      "isAllFree": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "요청 값 검증 실패 (INVALID_INPUT)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "INVALID_INPUT",
                  "message": "입력값이 올바르지 않습니다.",
                  "errors": [
                    {
                      "field": "token",
                      "message": "필수 값입니다."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "AUTH_SOCIAL_TOKEN_EXPIRED — 소셜 토큰 만료 · AUTH_SOCIAL_TOKEN_INVALID — 그 외 소셜 토큰 무효",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_SOCIAL_TOKEN_EXPIRED",
                  "message": "소셜 로그인 토큰이 만료되었습니다. 다시 로그인해 주세요."
                }
              }
            }
          },
          "503": {
            "description": "AUTH_SOCIAL_PROVIDER_UNAVAILABLE — 소셜 provider API 접근 실패(네트워크·타임아웃)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_SOCIAL_PROVIDER_UNAVAILABLE",
                  "message": "소셜 로그인 서버에 일시적으로 연결할 수 없습니다. 잠시 후 다시 시도해 주세요."
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "현재 사용자 조회",
        "description": "목적: 로그인 사용자 요약을 조회한다.\n\n호출 시점: 앱 진입·프로필/일정 변경 후 동기화.\n\n결과: UserSummary. hasPreSchedule은 일정 row 존재 여부(파생), isAllFree는 DB 컬럼.\n",
        "operationId": "me",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseUserSummaryResponse"
                },
                "example": {
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "email": "user@example.com",
                    "firstName": "길동",
                    "lastName": "홍",
                    "nickname": "홍길동",
                    "profileImageUrl": "https://lh3.googleusercontent.com/a/example",
                    "provider": "GOOGLE",
                    "isGoogleCalendarConnected": false,
                    "hasPreSchedule": false,
                    "isAllFree": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "API 에러 응답 envelope",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드",
            "example": "AUTH_INVALID_TOKEN"
          },
          "message": {
            "type": "string",
            "description": "사용자용 에러 메시지",
            "example": "유효하지 않은 토큰입니다."
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "description": "필드 검증 오류 목록 (validation 실패 시)",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "description": "요청 필드 검증 오류",
        "properties": {
          "field": {
            "type": "string",
            "description": "검증 실패 필드명",
            "example": "token"
          },
          "message": {
            "type": "string",
            "description": "검증 실패 메시지",
            "example": "must not be blank"
          }
        }
      },
...
      "SuccessResponseUserSummaryResponse": {
        "type": "object",
        "description": "API 성공 응답 envelope",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserSummaryResponse",
            "description": "응답 본문"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 메시지 (성공 시 null)"
          },
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 코드 (성공 시 null)"
          }
        }
      },
      "UserSummaryResponse": {
        "type": "object",
        "description": "사용자 요약. POST /auth/login · GET /auth/me · PATCH /users/onboarding/name · PATCH /users/profile 응답 공통",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "TripFit 사용자 ID (UUID v4)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "description": "소셜 계정 이메일. provider가 제공하지 않으면 null",
            "example": "user@example.com"
          },
          "firstName": {
            "type": [
              "string",
              "null"
            ],
            "description": "사용자가 입력한 이름. 미입력 시 null",
            "example": "길동"
          },
          "lastName": {
            "type": [
              "string",
              "null"
            ],
            "description": "사용자가 입력한 성. 미입력 시 null",
            "example": "홍"
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ],
            "description": "소셜 provider 표시명. 프로필 입력 prefill용. 미제공 시 null",
            "example": "홍길동"
          },
          "profileImageUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "프로필 이미지 URL. 현재는 소셜 provider CDN URL. 미제공 시 null",
            "example": "https://lh3.googleusercontent.com/a/example"
          },
          "provider": {
            "type": "string",
            "description": "로그인에 사용한 소셜 제공자",
            "enum": [
              "KAKAO",
              "GOOGLE",
              "APPLE"
            ]
          },
          "isGoogleCalendarConnected": {
            "type": "boolean",
            "description": "Google Calendar OAuth 연동 여부",
            "example": false
          },
          "hasPreSchedule": {
            "type": "boolean",
            "description": "정기 또는 개별 일정이 1건 이상 있는지 (DB 컬럼 없음, 조회 시 계산).\ntrue: 정기 일정 첫 생성 또는 개별 일정 첫 저장.\nfalse: 두 종류 일정 row가 모두 0건.\n일정 CRUD 응답에는 미포함 — GET /auth/me 등 재호출.\n",
            "example": false
          },
          "isAllFree": {
            "type": "boolean",
            "description": "전부 free 선언 여부 (user.is_all_free 저장, 기본 false).\n방 입장 조건: hasPreSchedule OR isAllFree.\ntrue: 일정 없이 Skip+0행으로 create/join·confirm 시 설정.\nfalse: 사용자가 명시적으로 해제하거나, 일정 row가 생기면 hasPreSchedule으로 대체 가능.\n일정 CRUD 응답에는 미포함 — GET /auth/me 등 재호출.\n",
            "example": false
          }
        }
      },
...
      "RefreshRequest": {
        "type": "object",
        "description": "액세스 토큰 재발급 요청. POST /auth/refresh",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "login 또는 이전 refresh 응답에서 받은 refresh token",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "minLength": 1
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RefreshResponse": {
        "type": "object",
        "description": "액세스 토큰 재발급 성공 응답. POST /auth/refresh",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "새 TripFit access JWT",
            "example": "eyJhbG..."
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64",
            "description": "access JWT 만료까지 남은 시간(초)",
            "example": 7200
          }
        }
      },
      "SuccessResponseRefreshResponse": {
        "type": "object",
        "description": "API 성공 응답 envelope",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RefreshResponse",
            "description": "응답 본문"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 메시지 (성공 시 null)"
          },
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 코드 (성공 시 null)"
          }
        }
      },
      "LogoutRequest": {
        "type": "object",
        "description": "로그아웃 요청. POST /auth/logout",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "서버에서 폐기할 refresh token",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "minLength": 1
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "description": "소셜 로그인 요청. POST /auth/login",
        "properties": {
          "provider": {
            "type": "string",
            "description": "소셜 로그인 제공자",
            "enum": [
              "KAKAO",
              "GOOGLE",
              "APPLE"
            ],
            "example": "GOOGLE"
          },
          "token": {
            "type": "string",
            "description": "소셜 토큰. GOOGLE/APPLE: id_token, KAKAO: access_token",
            "example": "eyJhbG...",
            "minLength": 1
          },
          "authorizationCode": {
            "type": [
              "string",
              "null"
            ],
            "description": "APPLE 전용 authorization code. 탈퇴 시 Apple revoke 호출(#64)에 필요한 refresh token 교환용 — GOOGLE/KAKAO는 보내지 않음",
            "example": "c1234..."
          }
        },
        "required": [
          "provider",
          "token"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "description": "소셜 로그인 성공 응답. POST /auth/login",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "TripFit API 호출용 access JWT",
            "example": "eyJhbG..."
          },
          "refreshToken": {
            "type": "string",
            "description": "access 만료 시 POST /auth/refresh에 사용하는 opaque refresh token",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64",
            "description": "access JWT 만료까지 남은 시간(초)",
            "example": 7200
          },
          "user": {
            "$ref": "#/components/schemas/UserSummaryResponse",
            "description": "로그인한 사용자 요약 (hasPreSchedule·isAllFree 포함)"
          }
        }
      },
      "SuccessResponseLoginResponse": {
        "type": "object",
        "description": "API 성공 응답 envelope",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LoginResponse",
            "description": "응답 본문"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 메시지 (성공 시 null)"
          },
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 코드 (성공 시 null)"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "type": "http",
        "description": "Authorization: Bearer {accessToken}",
        "name": "bearer-jwt",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}